+++ src/processors/modals/quote-maker.ts
@@ -4,7 +4,7 @@ import LocalQuotes from '../../main';
 import { getRandomQuoteId } from '../../utils/random';
 import { createDomLink } from '../../utils/dom';
 import { fetchAuthorsInQuoteVault, getValidAuthorsFromAdvancedSearch } from '../../types/quote';
-import { parseBlockMetadataToCodeBlock, parseTime } from '../../utils/parser';
+import {getAllAuthors, parseBlockMetadataToCodeBlock, parseTime} from '../../utils/parser';
 
 export class QuoteMakerModal extends Modal {
 	result: BlockMetadata;
@@ -94,18 +94,16 @@ export class QuoteMakerModal extends Modal {
 		.setButtonText('Check')
 		.setCta()
 		.onClick(() => {
+			let validatedAuthors: string[];
+
 			if (tmpSearch === '*') {
-				new Notice('You chose all authors.');
-				return;
+				validatedAuthors = getAllAuthors(this.plugin.settings.quoteVault);
+			} else {
+				validatedAuthors = getValidAuthorsFromAdvancedSearch(this.plugin.settings.quoteVault, tmpSearch);
 			}
 
-			const validatedAuthors = getValidAuthorsFromAdvancedSearch(
-				this.plugin.settings.quoteVault,
-				tmpSearch
-			).join('\n');
-
-			new Notice((validatedAuthors.length > 0)
-				? ('Validated authors:\n' + validatedAuthors)
+			new Notice(validatedAuthors.length > 0
+				? 'Validated authors:\n' + validatedAuthors.join('\n')
 				: 'Plugin can\'t find authors those you mentioned!\nTry to change search.'
 			);
 		}));