+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.4.0]
 ## Changed
 - Ability to use weighted random strategy in settings ([#15](https://github.com/ka1tzyu/local-quotes/issues/15))